requirements.txt
fastapi
bs4
requests
pymongo
python-decouple
uvicorn
selenium